+++ src/fn_native.rs
@@ -55,20 +55,19 @@ pub type Locked<T> = crate::stdlib::sync::RwLock<T>;
 
 /// Context of a native Rust function call.
 #[derive(Debug, Copy, Clone)]
-pub struct NativeCallContext<'e, 'n, 's, 'a, 'm> {
-    engine: &'e Engine,
-    fn_name: &'n str,
-    source: Option<&'s str>,
-    pub(crate) mods: Option<&'a Imports>,
-    pub(crate) lib: &'m [&'m Module],
+pub struct NativeCallContext<'a> {
+    engine: &'a Engine,
+    fn_name: &'a str,
+    source: Option<&'a str>,
+    mods: Option<&'a Imports>,
+    lib: &'a [&'a Module],
 }
 
-impl<'e, 'n, 's, 'a, 'm, M: AsRef<[&'m Module]> + ?Sized>
-    From<(&'e Engine, &'n str, Option<&'s str>, &'a Imports, &'m M)>
-    for NativeCallContext<'e, 'n, 's, 'a, 'm>
+impl<'a, M: AsRef<[&'a Module]> + ?Sized>
+    From<(&'a Engine, &'a str, Option<&'a str>, &'a Imports, &'a M)> for NativeCallContext<'a>
 {
     #[inline(always)]
-    fn from(value: (&'e Engine, &'n str, Option<&'s str>, &'a Imports, &'m M)) -> Self {
+    fn from(value: (&'a Engine, &'a str, Option<&'a str>, &'a Imports, &'a M)) -> Self {
         Self {
             engine: value.0,
             fn_name: value.1,
@@ -79,11 +78,11 @@ impl<'e, 'n, 's, 'a, 'm, M: AsRef<[&'m Module]> + ?Sized>
     }
 }
 
-impl<'e, 'n, 'm, M: AsRef<[&'m Module]> + ?Sized> From<(&'e Engine, &'n str, &'m M)>
-    for NativeCallContext<'e, 'n, '_, '_, 'm>
+impl<'a, M: AsRef<[&'a Module]> + ?Sized> From<(&'a Engine, &'a str, &'a M)>
+    for NativeCallContext<'a>
 {
     #[inline(always)]
-    fn from(value: (&'e Engine, &'n str, &'m M)) -> Self {
+    fn from(value: (&'a Engine, &'a str, &'a M)) -> Self {
         Self {
             engine: value.0,
             fn_name: value.1,
@@ -94,10 +93,10 @@ impl<'e, 'n, 'm, M: AsRef<[&'m Module]> + ?Sized> From<(&'e Engine, &'n str, &'m
     }
 }
 
-impl<'e, 'n, 's, 'a, 'm> NativeCallContext<'e, 'n, 's, 'a, 'm> {
+impl<'a> NativeCallContext<'a> {
     /// Create a new [`NativeCallContext`].
     #[inline(always)]
-    pub fn new(engine: &'e Engine, fn_name: &'n str, lib: &'m [&'m Module]) -> Self {
+    pub fn new(engine: &'a Engine, fn_name: &'a str, lib: &'a [&Module]) -> Self {
         Self {
             engine,
             fn_name,
@@ -112,11 +111,11 @@ impl<'e, 'n, 's, 'a, 'm> NativeCallContext<'e, 'n, 's, 'a, 'm> {
     #[cfg(not(feature = "no_module"))]
     #[inline(always)]
     pub fn new_with_all_fields(
-        engine: &'e Engine,
-        fn_name: &'n str,
-        source: &'s Option<&str>,
-        imports: &'a mut Imports,
-        lib: &'m [&'m Module],
+        engine: &'a Engine,
+        fn_name: &'a str,
+        source: &'a Option<&str>,
+        imports: &'a Imports,
+        lib: &'a [&Module],
     ) -> Self {
         Self {
             engine,
@@ -146,6 +145,14 @@ impl<'e, 'n, 's, 'a, 'm> NativeCallContext<'e, 'n, 's, 'a, 'm> {
     #[inline(always)]
     pub fn iter_imports(&self) -> impl Iterator<Item = (&str, &Module)> {
         self.mods.iter().flat_map(|&m| m.iter())
+    }
+    /// Get an iterator over the current set of modules imported via `import` statements.
+    #[cfg(not(feature = "no_module"))]
+    #[inline(always)]
+    pub(crate) fn iter_imports_raw(
+        &self,
+    ) -> impl Iterator<Item = (&ImmutableString, &Shared<Module>)> {
+        self.mods.iter().flat_map(|&m| m.iter_raw())
     }
     /// _(INTERNALS)_ The current set of modules imported via `import` statements.
     /// Available under the `internals` feature only.